requirements.txt
pillow
pandas
opencv_contrib_python==3.4.5.20
flask
flask-socketio
torchvision==0.2.1
reportlab==3.5.67
fpdf==1.7.2
keras==2.6.0
tensorflow==1.14.0
gunicorn==20.0.4
flask==1.0.3
seaborn
PyPDF2==1.26.0
scikit-image
scikit-learn
imutils
wxpython
